further changes on unify_bad_channels
