+++ mne/preprocessing/bads.py
@@ -1,18 +1,9 @@
 # Authors: Denis Engemann <denis.engemann@gmail.com>
-#          Ana Radanovic <radanovica@protonmail.com>
-#          Erica Peterson <nordme@uw.edu>
+#
 # License: BSD-3-Clause
 
 import numpy as np
 from scipy.stats import zscore
-
-
-from ..utils import _validate_type
-
-from ..io import BaseRaw
-from ..epochs import Epochs
-from ..evoked import Evoked
-from ..time_frequency.spectrum import BaseSpectrum
 
 
 def _find_outliers(X, threshold=3.0, max_iter=2, tail=0):
@@ -57,71 +48,3 @@ def _find_outliers(X, threshold=3.0, max_iter=2, tail=0):
 
     bad_idx = np.where(my_mask)[0]
     return bad_idx
-
-
-def unify_bad_channels(insts):
-    """Unify bad channels across a list of instances, using the union.
-
-    Parameters
-    ----------
-    insts : list
-        List of instances (:class:`~mne.io.Raw`, :class:`~mne.Epochs`,
-        :class:`~mne.Evoked`, :class:`~mne.time_frequency.Spectrum`,
-        :class:`~mne.time_frequency.EpochSpectrum`) across which to unify bad channels.
-
-    Returns
-    -------
-    insts : list
-        List of instances with bad channels unified across instances.
-
-    Notes
-    -----
-    This function operates in-place.
-
-    .. versionadded:: 1.6
-    """
-    # check input is not an empty list
-    if len(insts) == 0:
-        raise ValueError("Be sure insts is not empty list")
-
-    # first check that each object is mne object
-    inst_type = type(insts[0])
-    valid_types = (BaseRaw, Epochs, Evoked, BaseSpectrum)
-    for inst in insts:
-        _validate_type(inst, valid_types, "instance type")
-        if type(inst) != inst_type:
-            raise ValueError("All insts must be the same type")
-
-    # check that input is a list
-    if not isinstance(insts, list):
-        raise ValueError(f"insts must be a *list* of mne objects, got {type(insts)}")
-
-    # check that all channels have the same name and same number
-    ch_names = insts[0].ch_names
-    diff_chns = []
-    for inst in insts[1:]:
-        if inst.ch_names != ch_names:
-            dif = set(inst.info.ch_names) ^ (set(ch_names))
-            diff_chns.extend(list(dif))
-
-    if len(diff_chns) > 0:
-        raise ValueError(
-            "Channel names are not consistent across instances."
-            f" Mismatch channels are {diff_chns}"
-        )
-
-    # collect bads as dict keys so that insertion order is preserved
-    # then later cast to list.
-    all_bads = dict()
-    for inst in insts:
-        all_bads.update(dict.fromkeys(inst.info["bads"]))
-    all_bads = list(all_bads)
-
-    # apply bads set to all instances
-    new_insts = []
-
-    for inst in insts:
-        inst.info["bads"] = all_bads
-        new_insts.append(inst)
-
-    return new_insts
